+++ ChitChatClient/Windows/ChatWindow.xaml.cs
@@ -1,4 +1,5 @@
-﻿using ChitChatClient.Models;
+﻿using ChitChatClient.Helpers;
+using ChitChatClient.Models;
 using ChitChatClient.Services;
 using ChitChatClient.ViewModels;
 using System;
@@ -30,19 +31,6 @@ namespace ChitChatClient.Windows
 
 		private readonly ServerService serverService = ServerService.Instance;
 
-		public ChatWindow()
-		{
-			InitializeComponent();
-			DataContext = chatInterface;
-
-			// Initializing user refresh timer
-			userRefreshTimer = new DispatcherTimer();
-
-			userRefreshTimer.Tick += new EventHandler(RefreshServerUsersTask);
-			userRefreshTimer.Interval = new TimeSpan(0, 0, 5);
-			userRefreshTimer.Start();
-		}
-
 		private Server? SelectedServer
 		{
 			get
@@ -59,6 +47,49 @@ namespace ChitChatClient.Windows
 			}
 		}
 
+		private Channel? SelectedChannel
+		{
+			get
+			{
+				Channel selectedChannel = (Channel)channelListView.SelectedItem;
+				if (selectedChannel == null)
+				{
+					return null;
+				}
+				else
+				{
+					return selectedChannel;
+				}
+			}
+		}
+
+		public ChatWindow()
+		{
+			InitializeComponent();
+			DataContext = chatInterface;
+
+			// Initializing user refresh timer
+			userRefreshTimer = new DispatcherTimer();
+
+			userRefreshTimer.Tick += new EventHandler(RefreshServerUsersTask);
+			userRefreshTimer.Interval = new TimeSpan(0, 0, 5);
+			userRefreshTimer.Start();
+
+			// Set placeholders
+			PlaceholderProperty.SetPlaceholderText(userFilterTextBox);
+			PlaceholderProperty.SetPlaceholderText(filterServerTextBox);
+			PlaceholderProperty.SetPlaceholderText(messageTextBox);
+		}
+
+		private async void Window_Loaded(object sender, RoutedEventArgs e)
+		{
+			await chatInterface.RefreshJoinedServers(RefreshSeverListView);
+			await chatInterface.RefreshServerUsers(RefreshUserListView, SelectedServer);
+			await RefreshServerChannels();
+			CollectionViewSource.GetDefaultView(userListView.ItemsSource).Filter = FilterUser;
+			CollectionViewSource.GetDefaultView(serverListView.ItemsSource).Filter = FilterServer;
+		}
+
 		private void RefreshSeverListView()
 		{
 			serverListView.GetBindingExpression(ListView.ItemsSourceProperty).UpdateTarget();
@@ -71,7 +102,6 @@ namespace ChitChatClient.Windows
 
 		private void RefreshUserListView(int? onlineUserCount, int? totalUserCount)
 		{
-			userFilterTextBox.Text = string.Empty;
 			userListView.GetBindingExpression(ListView.ItemsSourceProperty).UpdateTarget();
 			CollectionViewSource.GetDefaultView(userListView.ItemsSource).Filter = FilterUser;
 			CollectionViewSource.GetDefaultView(userListView.ItemsSource).Refresh();
@@ -94,12 +124,16 @@ namespace ChitChatClient.Windows
 			await RefreshServerUsersWrapper();
 		}
 
-        private async void Window_Loaded(object sender, RoutedEventArgs e)
-        {
-            await chatInterface.RefreshJoinedServers(RefreshSeverListView);
-            await chatInterface.RefreshServerUsers(RefreshUserListView, SelectedServer);
-			CollectionViewSource.GetDefaultView(userListView.ItemsSource).Filter = FilterUser;
-			CollectionViewSource.GetDefaultView(serverListView.ItemsSource).Filter = FilterServer;
+		private async Task RefreshServerChannels()
+		{
+			await chatInterface.GetServerChannels(SelectedServer);
+
+			if (SelectedChannel != null)
+			{
+				currentChannelLabel.Text = SelectedChannel.Name;
+				currentChannelDescriptionLabel.Text = SelectedChannel.Description;
+			}
+
 		}
 
 		private async void createServerButton_Click(object sender, RoutedEventArgs e)
@@ -147,7 +181,8 @@ namespace ChitChatClient.Windows
 
         private void serverSettingsButton_Click(object sender, RoutedEventArgs e)
 		{
-            MessageBox.Show("Feature not implemented", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
+			ServerSettingsWindow serverSettingsWindow = new();
+			serverSettingsWindow.ShowDialog();
         }
 
 		private void sendButton_Click(object sender, RoutedEventArgs e)
@@ -168,7 +203,8 @@ namespace ChitChatClient.Windows
 		private async void serverListView_SelectionChanged(object sender, SelectionChangedEventArgs e)
 		{
 			await chatInterface.RefreshServerUsers(RefreshUserListView, SelectedServer);
-        }
+			await RefreshServerChannels();
+		}
 
 		private void Window_Closing(object sender, System.ComponentModel.CancelEventArgs e)
 		{
@@ -177,12 +213,18 @@ namespace ChitChatClient.Windows
 
 		private void userFilterTextBox_TextChanged(object sender, TextChangedEventArgs e)
 		{
-			CollectionViewSource.GetDefaultView(userListView.ItemsSource).Refresh();
+			if (userListView.ItemsSource != null)
+			{
+				CollectionViewSource.GetDefaultView(userListView.ItemsSource).Refresh();
+			}
 		}
 
 		private void filterServerTextBox_TextChanged(object sender, TextChangedEventArgs e)
 		{
-			CollectionViewSource.GetDefaultView(serverListView.ItemsSource).Refresh();
+			if (serverListView.ItemsSource != null)
+			{
+				CollectionViewSource.GetDefaultView(serverListView.ItemsSource).Refresh();
+			}
 		}
 
 		private bool FilterUser(object obj)
@@ -192,9 +234,15 @@ namespace ChitChatClient.Windows
 				return true;
 			}
 
+			// If the text is the placeholder, return true
+			if (userFilterTextBox.Text == PlaceholderProperty.GetPlaceholder(userFilterTextBox))
+			{
+				return true;
+			}
+
 			if (obj is User user)
 			{
-				if (user.Username.ToLower().Contains(userFilterTextBox.Text.ToLower()))
+				if (user.Username.Contains(userFilterTextBox.Text, StringComparison.CurrentCultureIgnoreCase))
 				{
 					return true;
 				}
@@ -210,15 +258,51 @@ namespace ChitChatClient.Windows
 				return true;
 			}
 
+			// If the text is the placeholder, return true
+			if (filterServerTextBox.Text == PlaceholderProperty.GetPlaceholder(filterServerTextBox))
+			{
+				return true;
+			}
+
 			if (obj is Server server)
 			{
-				if (server.Name.ToLower().Contains(filterServerTextBox.Text.ToLower()))
+				if (server.Name.Contains(filterServerTextBox.Text, StringComparison.CurrentCultureIgnoreCase))
 				{
 					return true;
 				}
 			}
 
 			return false;
+		}
+
+		private void userFilterTextBox_GotFocus(object sender, RoutedEventArgs e)
+		{
+			PlaceholderProperty.ClearPlaceholderText(userFilterTextBox);
+		}
+
+		private void userFilterTextBox_LostFocus(object sender, RoutedEventArgs e)
+		{
+			PlaceholderProperty.SetPlaceholderText(userFilterTextBox);
+		}
+
+		private void filterServerTextBox_GotFocus(object sender, RoutedEventArgs e)
+		{
+			PlaceholderProperty.ClearPlaceholderText(filterServerTextBox);
+		}
+
+		private void filterServerTextBox_LostFocus(object sender, RoutedEventArgs e)
+		{
+			PlaceholderProperty.SetPlaceholderText(filterServerTextBox);
+		}
+
+		private void messageTextBox_GotFocus(object sender, RoutedEventArgs e)
+		{
+			PlaceholderProperty.ClearPlaceholderText(messageTextBox);
+		}
+
+		private void messageTextBox_LostFocus(object sender, RoutedEventArgs e)
+		{
+			PlaceholderProperty.SetPlaceholderText(messageTextBox);
 		}
 	}
 }