+++ release_notes.md
@@ -1,8 +1,12 @@
+## Update your launcher, new version is out!
+
 ## New features
 
 - Added checksum checking to launcher
 	- It will update if there is a new version
 	- It will update if previous update failed
+- Implemented a server settings window
+- Implemented placeholders for filter fields
 
 ## Changes